move (clipping) plane slider business to its own depiction class for easier reuse in other layers
